More AMD JS updates
