+++ src/twist/Core/Resources/twist/ajax/css/twistajax.css
@@ -62,4 +62,4 @@
   #twist-ajax-loader.small #twist-ajax-loader-size, #twist-ajax-loader.smallest #twist-ajax-loader-count, #twist-ajax-loader.smallest #twist-ajax-loader-size {
     display: none; }
 
-/*# sourceMappingURL=twist-ajax.css.map */
+/*# sourceMappingURL=twistajax.css.map */
